requirements.txt
aiohttp==3.8.2
pytest-asyncio==0.21.1